+++ src/OpenMonday.Core/MondayDriver/Services/Interfaces/IMondayBoardDriverService.cs
@@ -5,4 +5,6 @@ public interface IMondayBoardDriverService
     Task<MondayDriverResult<MondayDriverBoardStructure>> GetBoardsStructureById(string boardId);
 
     Task<MondayDriverResult<List<MondayDriverBaseTask>>> GetBoardItemsByCursor(string board_id);
+
+    Task<MondayDriverResult<List<MondayDriverActivityLog>>> GetActivityLogs(string board_id, DateTime from, DateTime to);
 }